// src/interfaces/provider.interface.ts
export interface Provider  {
    businessName: string;
    contactName: string;
    phone: string;
    email: string;
    address: string;
    merchandiseType: string;
    type: string;
    description: string;
    rut: string;
    web: string;
    instagran: string;
    facebook: string;
    catalogue: string;
    created: Date;
}